♻️ [Refactoring]: #304 [FE] パス末尾セグメント抽出を utils/path の basenameOf に集約
useRecentProjects 内に閉じていた basenameOf を汎用ユーティリティ
src/utils/path へ移し、App.tsx のインライン末尾セグメント抽出も
これに委譲する。両実装は末尾の空セグメントを除外する点で等価なため
挙動は不変。basenameOf のテストも utils/path 側へ移設する。
