+++ src/hooks/useRecentProjects/__tests__/useRecentProjects.unit.test.ts
@@ -1,19 +1,9 @@
 import { expect, test } from "vitest";
 import {
   addRecentProject,
-  basenameOf,
   normalizeRecentProjects,
   type RecentProject,
 } from "../index";
-
-test.each([
-  ["/home/user/my-project", "my-project"],
-  ["/home/user/my-project/", "my-project"],
-  ["C:\\Users\\me\\proj", "proj"],
-  ["single", "single"],
-])("basenameOf(%s) は末尾セグメントを返す", (input, expected) => {
-  expect(basenameOf(input)).toBe(expected);
-});
 
 test("addRecentProject は新規パスを先頭に name 付きで追加する", () => {
   const result = addRecentProject([], "/home/user/proj");
